+++ LibraryCPP/queue.h
@@ -3,7 +3,7 @@
 
 // Queue
 // Stores integer values inside
-typedef size_t Data;
+typedef int Data;
 
 struct Queue;
